Write a function to encrypt API key in utils
We need a function that encrypts api keys before storing them.
API keys are sensitive and should be stored with the utmost care.

Considering the fact that there is no trusting the server, and present
field encryption mechanisms are either not suitable or hard to
implement right now, a set of encryption/decryption functions have been
created for use.

Also, a mechanism to rotate keys on exit has been implemented.